requirement.txt
absl-py==0.7.1
albumentations==0.2.3
astor==0.7.1
certifi==2019.6.16
cffi==1.12.3
cycler==0.10.0
decorator==4.4.0
future==0.17.1
grpcio==1.21.1
imageio==2.5.0
imgaug==0.2.6
joblib==0.13.2
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.0
networkx==2.3
numpy==1.16.4
olefile==0.46
opencv-python==4.1.0.25
opencv-python-headless==4.1.0.25
Pillow==6.0.0
protobuf==3.8.0
pycparser==2.19
pyparsing==2.4.0
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1.1
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
sklearn==0.0
tb-nightly==1.15.0a20190624
torch==1.1.0
torchvision==0.3.0
tqdm==4.33.0
Werkzeug==0.15.4
yacs==0.1.6
